@@ -1,34 +1,46 @@
-# Copyright 2023 Google LLC
-# 
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# 
-#      https://www.apache.org/licenses/LICENSE-2.0
-# 
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 [build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling", "hatch-requirements-txt"]
+build-backend = "hatchling.build"
+
+[tool.hatch.version]
+path = "src/MaxText/__init__.py"
 
 [project]
-name = "MaxText"
-version = "0.1.0"
-description = "MaxText is a high performance, highly scalable, open-source LLM written in pure Python/Jax and targeting Google Cloud TPUs and GPUs for training and inference"
+name = "maxtext"
+description = "MaxText is a simple, performant and scalable Jax LLM!"
+dynamic = ["version", "optional-dependencies"]
+requires-python = ">=3.12"
 readme = "README.md"
-requires-python = ">=3.10"
-license = {file = "LICENSE"}
-dynamic = ["dependencies"]
+license = "Apache-2.0"
+keywords = ["llm", "jax", "llama", "mistral", "mixtral", "gemma", "deepseek"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "Intended Audience :: Science/Research",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: ML",
+    "Environment :: GPU :: NVIDIA CUDA :: 12",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+]
+dependencies = []
 
-[tool.setuptools.dynamic]
-dependencies = {file = ["requirements.txt"]}
+[tool.hatch.metadata.hooks.requirements_txt.optional-dependencies]
+tpu = ["dependencies/requirements/generated_requirements/tpu-requirements.txt"]
+cuda12 = ["dependencies/requirements/generated_requirements/cuda12-requirements.txt"]
 
-[tool.setuptools.packages.find]  
-where = ["MaxText"]
-include = ["MaxText*"]
-exclude = ["MaxText.tests.*"]
+[project.urls]
+Repository = "https://github.com/AI-Hypercomputer/maxtext.git"
+"Bug Tracker" = "https://github.com/AI-Hypercomputer/maxtext/issues"
+
+[tool.hatch.metadata]
+allow-direct-references = true
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/MaxText", "src/install_maxtext_extra_deps"]
+
+[tool.hatch.build.targets.wheel.hooks.custom]
+path = "build_hooks.py"
+
+[project.scripts]
+install_maxtext_github_deps = "install_maxtext_extra_deps.install_github_deps:main"
