@@ -1,24 +1,29 @@
+# Copyright 2023–2025 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""This file contains utility functions for running shell commands.
+
+It provides helpers for:
+  - Running a single command with real-time output streaming.
+  - Running multiple commands in parallel batches.
+  - Creating temporary log files for command outputs.
 """
- Copyright 2024 Google LLC
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      https://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
- """
-
 
 import datetime
+import subprocess
 import sys
 import tempfile
-import subprocess
 import time
 
 
@@ -30,7 +35,7 @@ def chunks(lst: list, n: int):
     n: size of each chunk.
 
   Returns:
-    List of n-sized chunks for lst.
+    list of n-sized chunks for lst.
   """
   return [lst[i : i + n] for i in range(0, len(lst), n)]
 
@@ -46,10 +51,7 @@ def make_tmp_files(per_command_name):
   """
   # Supports removal of spaces from command names before converting to file name.
   return [
-      tempfile.NamedTemporaryFile(
-          delete=False, prefix=command.replace(' ', '-') + '-'
-      )
-      for command in per_command_name
+      tempfile.NamedTemporaryFile(delete=False, prefix=command.replace(" ", "-") + "-") for command in per_command_name
   ]
 
 
@@ -70,17 +72,14 @@ def run_commands(commands, jobname, per_command_name, batch=10, dry_run=False):
   commands_batched = chunks(commands, batch)
   per_command_name_batches = chunks(per_command_name, batch)
 
-  print(
-      f'Breaking up a total of {len(commands)} commands into'
-      f' {len(commands_batched)} batches'
-  )
+  print(f"Breaking up a total of {len(commands)} commands into" f" {len(commands_batched)} batches")
   if dry_run:
-    print('Pretending all the jobs succeeded')
+    print("Pretending all the jobs succeeded")
     return 0
 
   max_return_code = 0
   for i, _ in enumerate(commands_batched):
-    print(f'Dispatching batch {i}/{len(commands_batched)}')
+    print(f"Dispatching batch {i}/{len(commands_batched)}")
     batch_max_return_code, _ = run_command_batch(
         commands_batched[i],
         jobname,
@@ -111,9 +110,7 @@ def run_command_batch(commands, jobname, per_command_name, output_logs):
   for i, command in enumerate(commands):
     children.append(
         # subprocess managed by list pylint: disable=consider-using-with
-        subprocess.Popen(
-            command, stdout=output_logs[i], stderr=output_logs[i], shell=True
-        )
+        subprocess.Popen(command, stdout=output_logs[i], stderr=output_logs[i], shell=True)
     )
 
   while True:
@@ -125,25 +122,14 @@ def run_command_batch(commands, jobname, per_command_name, output_logs):
     if completed < total:
       slow_worker_index = returncodes.index(None)
       slow_worker_text = per_command_name[slow_worker_index]
-      slow_str = (
-          f', task {slow_worker_text} still working, logfile'
-          f' {output_logs[slow_worker_index].name}'
-      )
+      slow_str = f", task {slow_worker_text} still working, logfile" f" {output_logs[slow_worker_index].name}"
     else:
-      slow_str = ''
-    print(
-        f'[t={seconds_elapsed:.2f}, {jobname}] Completed'
-        f' {completed}/{total}{slow_str}'
-    )
+      slow_str = ""
+    print(f"[t={seconds_elapsed:.2f}, {jobname}] Completed" f" {completed}/{total}{slow_str}")
     if max_returncode > 0:
-      failing_index = [
-          i for i, x in enumerate(returncodes) if x is not None and x > 0
-      ][0]
-      print(f'Terminating all {jobname} processes since at least one failed.')
-      print(
-          f'Failure is {per_command_name[failing_index]}'
-          f' and logfile {output_logs[failing_index].name}'
-      )
+      failing_index = [i for i, x in enumerate(returncodes) if x is not None and x > 0][0]
+      print(f"Terminating all {jobname} processes since at least one failed.")
+      print(f"Failure is {per_command_name[failing_index]}" f" and logfile {output_logs[failing_index].name}")
       for child in children:
         child.terminate()
       break
@@ -170,9 +156,7 @@ def run_command_with_updates(command, task, verbose=True) -> int:
   """
 
   if verbose:
-    print(
-        f'Task: `{task}` is implemented by `{command}`, streaming output live.'
-    )
+    print(f"Task: `{task}` is implemented by `{command}`, streaming output live.")
     with subprocess.Popen(
         command,
         stdout=sys.stdout,
@@ -183,27 +167,21 @@ def run_command_with_updates(command, task, verbose=True) -> int:
       while True:
         return_code = child.poll()
         if return_code is None:
-          print(f'Waiting for `{task}`, for {i} seconds')
+          print(f"Waiting for `{task}`, for {i} seconds")
           time.sleep(1)
           i += 1
         else:
-          print(f'Task: `{task}` terminated with code `{return_code}`')
+          print(f"Task: `{task}` terminated with code `{return_code}`")
           return return_code
   else:
-    print(
-        f'Task: `{task}` is implemented by `{command}`, hiding output unless'
-        ' there is an error.'
-    )
+    print(f"Task: `{task}` is implemented by `{command}`, hiding output unless" " there is an error.")
     try:
       subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
     except subprocess.CalledProcessError as e:
-      print(
-          f'Task: `{task}` terminated with ERROR `{e.returncode}`, printing'
-          ' logs'
-      )
-      print('*' * 80)
+      print(f"Task: `{task}` terminated with ERROR `{e.returncode}`, printing" " logs")
+      print("*" * 80)
       print(e.output)
-      print('*' * 80)
+      print("*" * 80)
       return e.returncode
-    print(f'Task: `{task}` succeeded.')
+    print(f"Task: `{task}` succeeded.")
     return 0
