@@ -1,10 +1,10 @@
-# Copyright 2025 Google LLC
+# Copyright 2023–2025 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#    https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,9 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-# Dictionary mapping each subject to its respective subcategories.
-# Subcategories represent more specific topics which are later grouped into broader categories.
+"""
+Dictionary mapping each subject to its respective subcategories.
+Subcategories represent more specific topics which are later grouped into broader categories.
+"""
 
 subcategories = {
     # STEM Subjects
