@@ -1,20 +1,25 @@
-"""Copyright 2025 Google LLC
+# Copyright 2023–2025 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
 """
+This recipe demonstrates how to launch a Pathways workload in headless mode.
 
-import os
-import sys
+In headless mode, the XPK workload starts the Pathways server and proxy
+components but does not run a user command. This is useful for setting up a
+persistent training environment that can be connected to later by a separate
+runner process.
+"""
 
 import benchmarks.recipes.args_helper as helper
 import maxtext_xpk_runner as mxr
@@ -23,9 +28,7 @@ from recipes.user_configs import cluster_config, xpk_path, pathways_config, base
 
 def main() -> int:
   # Handle command line arguments using args_helper
-  should_continue = helper.handle_cmd_args(
-      cluster_config, helper.DELETE, xpk_path=xpk_path
-  )
+  should_continue = helper.handle_cmd_args(cluster_config, helper.DELETE, xpk_path=xpk_path)
 
   if not should_continue:
     return 0
@@ -57,6 +60,8 @@ def main() -> int:
   return_code = mxr.run_command_with_updates(command, name)
   if return_code != 0:
     print(f"Unable to run xpk workload: {name}")
+
+  return return_code
 
 
 if __name__ == "__main__":
