@@ -1,31 +1,34 @@
-"""Copyright 2025 Google LLC
+# Copyright 2023–2025 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+"""A recipe for running a MaxText benchmark using Pathways with a remote Python sidecar.
 
-    https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+This script configures and launches a workload on a GKE cluster using XPK.
+It defines the cluster, Docker images for the server, proxy, and runner,
+and sets up the model configuration for a Pathways-based run.
 """
 
 import os
-import sys
+
 import args_helper as helper
 
-parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
-sys.path.append(parent_dir)
-
-import maxtext_trillium_model_configs as model_configs
-import maxtext_xpk_runner as mxr
-from xpk_configs import XpkClusterConfig
+from benchmarks import maxtext_trillium_model_configs as model_configs
+from benchmarks import maxtext_xpk_runner as mxr
+from benchmarks.xpk_configs import XpkClusterConfig
 
 
-def main() -> int:
+def main():
   # V6e cluster config
   cluster_config = XpkClusterConfig(
       cluster_name="v6e-256-cluster",
@@ -37,24 +40,16 @@ def main() -> int:
   xpk_path = "xpk"
 
   # Handle command line arguments using args_helper
-  should_continue = helper.handle_cmd_args(
-      cluster_config, helper.DELETE, xpk_path=xpk_path
-  )
+  should_continue = helper.handle_cmd_args(cluster_config, helper.DELETE, xpk_path=xpk_path)
 
   if not should_continue:
-    return 0
+    return
 
   # Configure test images
   user = os.environ["USER"]
   region = "-".join(cluster_config.zone.split("-")[:-1])
-  proxy_image = (
-      f"us-docker.pkg.dev/cloud-tpu-v2-images/pathways/gke/{user}/"
-      "proxy_server:latest"
-  )
-  server_image = (
-      f"us-docker.pkg.dev/cloud-tpu-v2-images/pathways/gke/{user}/"
-      "server:latest"
-  )
+  proxy_image = f"us-docker.pkg.dev/cloud-tpu-v2-images/pathways/gke/{user}/" "proxy_server:latest"
+  server_image = f"us-docker.pkg.dev/cloud-tpu-v2-images/pathways/gke/{user}/" "server:latest"
   colocated_python_image = f"gcr.io/{cluster_config.project}/{user}/colocated_python_sidecar_latest:latest"
   runner = f"gcr.io/{cluster_config.project}/{user}_latest:latest"
   base_output_directory = f"gs://{user}-{region}/{user}"
@@ -93,9 +88,7 @@ def main() -> int:
             xpk_path=xpk_path,
             num_steps=1000000,
         )
-        command, name = mxr.generate_xpk_workload_cmd(
-            cluster_config=cluster_config, wl_config=wl_config
-        )
+        command, name = mxr.generate_xpk_workload_cmd(cluster_config=cluster_config, wl_config=wl_config)
 
         print(f"Name of the workload is: {name} \n")
         xpk_workload_names.append(name)
@@ -103,12 +96,8 @@ def main() -> int:
         print(f"XPK command to be used is: {command} \n")
         xpk_workload_cmds.append(command)
 
-  for xpk_workload_name, xpk_workload_cmd in zip(
-      xpk_workload_names, xpk_workload_cmds
-  ):
-    return_code = mxr.run_command_with_updates(
-        xpk_workload_cmd, xpk_workload_name
-    )
+  for xpk_workload_name, xpk_workload_cmd in zip(xpk_workload_names, xpk_workload_cmds):
+    return_code = mxr.run_command_with_updates(xpk_workload_cmd, xpk_workload_name)
     if return_code != 0:
       print(f"Unable to run xpk workload: {xpk_workload_name}")
 
