@@ -1,18 +1,16 @@
-"""
- Copyright 2024 Google LLC
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      https://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
- """
+# Copyright 2023–2025 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """This file contains commonly-used XLA Flags."""
 
 ### NOTICE ###
@@ -83,11 +81,9 @@ ENABLE_SPARSECORE_OFFLOADING_FOR_RS_AG_AR = (
     " --xla_tpu_enable_async_collective_fusion_fuse_all_gather=false"
     " --xla_tpu_enable_async_collective_fusion_fuse_all_reduce=false"
     " --xla_tpu_enable_async_collective_fusion_fuse_reduce_scatter=false"
-
     " --xla_tpu_enable_sparse_core_collective_offload_all_gather=true"
     " --xla_tpu_enable_sparse_core_collective_offload_reduce_scatter=true"
     " --xla_tpu_enable_sparse_core_collective_offload_all_reduce=true"
-
     " --xla_tpu_enable_all_gather_offload_tracing=true"
     " --xla_tpu_enable_reduce_scatter_offload_tracing=true"
     " --xla_tpu_enable_all_reduce_offload_tracing=true"
@@ -138,16 +134,13 @@ REDUCE_SCATTER_FUSION = (
 # xla_tpu_enable_data_parallel_all_reduce_opt:
 #   optimize DCN all-reduces used for data parallel sharding
 DATA_PARALLEL_OVERLAP = (
-    " --xla_tpu_enable_data_parallel_all_reduce_opt=true"
-    " --xla_tpu_data_parallel_opt_different_sized_ops=true"
+    " --xla_tpu_enable_data_parallel_all_reduce_opt=true" " --xla_tpu_data_parallel_opt_different_sized_ops=true"
 )
 
 # Enable Enhanced Launch Barrier.
 # Gracefully handle dispatch failures on TPU workers. For Pathways is required
 # for error propagation and out-of-order execution detection.
-ENHANCED_LAUNCH_BARRIER = (
-    " --xla_tpu_use_enhanced_launch_barrier=true"
-)
+ENHANCED_LAUNCH_BARRIER = " --xla_tpu_use_enhanced_launch_barrier=true"
 
 # Host offloading Flags. These are optimizations recommended when using host
 # offloading.
@@ -166,15 +159,11 @@ HOST_OFFLOAD_FLAGS = (
 )
 
 # Flags to optimize pipeline parallelism over DCN with large host offloading.
-PIPELINING_FLAGS = (
-    " --xla_tpu_iova_dma_chunk_size_bytes=16777216" # breaks DMA to/from host into 16M chunks
-)
+PIPELINING_FLAGS = " --xla_tpu_iova_dma_chunk_size_bytes=16777216"  # breaks DMA to/from host into 16M chunks
 
 # Disable bundle-aware CostModel which was causing worse perf b/357103386.
 # Some fusions in the backward pass of the model were 3x slower without this.
-DISABLE_BUNDLE_AWARE_COST_MODEL = (
-    " --xla_tpu_use_bundle_aware_cost_model_for_fusions=false"
-)
+DISABLE_BUNDLE_AWARE_COST_MODEL = " --xla_tpu_use_bundle_aware_cost_model_for_fusions=false"
 
 # Enable Silent Data Corruption (SDC) Checker
 # SDC Checker will check for chip / ici / hardware corruption events.
@@ -201,6 +190,8 @@ DEBUG_LOGS = {
 }
 
 # Disables collective matmul operations.
-DISABLE_COLLECTIVE_MATMUL = (
-    " --xla_jf_spmd_threshold_for_windowed_einsum_mib=1000000"
-)
+DISABLE_COLLECTIVE_MATMUL = " --xla_jf_spmd_threshold_for_windowed_einsum_mib=1000000"
+# Disable "megacore fusion allow ags".
+DISABLE_MEGACORE_FUSION_ALLOW_AGS = " --xla_tpu_megacore_fusion_allow_ags=false"
+# Enable async collective permute.
+ENABLE_ASYNC_COLLECTIVE_PERMUTE = " --xla_enable_async_collective_permute=true"
