@@ -1,16 +1,17 @@
 # pylint: skip-file
-"""
- Copyright 2025 Google LLC
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-      https://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
-"""
+# Copyright 2023–2025 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """Test module for XLA GPU feature verification.
 
 This module provides test utilities for verifying GPU-specific features in HLO files,
@@ -18,7 +19,7 @@ including collective matmul operations and FP8 GEMM operations.
 """
 
 from absl import app
-from typing import Sequence, Tuple
+from typing import Sequence
 
 import numpy as np
 import re
@@ -50,8 +51,8 @@ class XlaGpuFeatureTestCase:
     expected_unrolled_ag = int(expected_unrolled_ag)
     expected_unrolled_rs = int(expected_unrolled_rs)
 
-    pattern_ag = r'%unrolled_windowed_dot_general_body_ag'
-    pattern_rs = r'%unrolled_windowed_dot_general_body_rs'
+    pattern_ag = r"%unrolled_windowed_dot_general_body_ag"
+    pattern_rs = r"%unrolled_windowed_dot_general_body_rs"
 
     actual_unrolled_ag = len(re.findall(pattern_ag, hlo_content, re.MULTILINE))
     actual_unrolled_rs = len(re.findall(pattern_rs, hlo_content, re.MULTILINE))
@@ -88,18 +89,14 @@ class XlaGpuFeatureTestCase:
 
     expected_fp8_gemm = int(expected_fp8_gemm)
 
-    pattern_fp8_gemm = r'__cublas\$lt\$matmul\$f8'
+    pattern_fp8_gemm = r"__cublas\$lt\$matmul\$f8"
 
     actual_fp8_gemm = len(re.findall(pattern_fp8_gemm, hlo_content, re.MULTILINE))
 
     np.testing.assert_equal(
         actual_fp8_gemm,
         expected_fp8_gemm,
-        err_msg=(
-            f"FP8 GEMM operation count mismatch. "
-            f"Expected: {expected_fp8_gemm}, "
-            f"Actual: {actual_fp8_gemm}"
-        ),
+        err_msg=(f"FP8 GEMM operation count mismatch. " f"Expected: {expected_fp8_gemm}, " f"Actual: {actual_fp8_gemm}"),
     )
 
 
@@ -117,10 +114,10 @@ def test_collective_matmul(
     expected_unrolled_ag: Expected number of unrolled all-gather operations.
     expected_unrolled_rs: Expected number of unrolled reduce-scatter operations.
   """
-  with open(hlo_file, 'r') as hlo_file:
+  with open(hlo_file, "r") as hlo_file:
     hlo_content = hlo_file.read()
     test_case.check_collective_matmul(hlo_content, expected_unrolled_ag, expected_unrolled_rs)
-    print('Collective matmul test passed.')
+    print("Collective matmul test passed.")
 
 
 def test_fp8_gemm(
@@ -135,10 +132,10 @@ def test_fp8_gemm(
     hlo_file: Path to the HLO file.
     expected_fp8_gemm: Expected number of FP8 GEMM operations.
   """
-  with open(hlo_file, 'r') as hlo_file:
+  with open(hlo_file, "r") as hlo_file:
     hlo_content = hlo_file.read()
     test_case.check_fp8_gemm(hlo_content, expected_fp8_gemm)
-    print('FP8 GEMM test passed.')
+    print("FP8 GEMM test passed.")
 
 
 def main(argv: Sequence[str]) -> None:
@@ -154,9 +151,9 @@ def main(argv: Sequence[str]) -> None:
   test_case = XlaGpuFeatureTestCase()
   _, test_scenario, *test_vars = argv
 
-  if test_scenario == 'collective_matmul':
+  if test_scenario == "collective_matmul":
     test_collective_matmul(test_case, *test_vars)
-  elif test_scenario == 'fp8_gemm':
+  elif test_scenario == "fp8_gemm":
     test_fp8_gemm(test_case, *test_vars)
   else:
     raise ValueError(f"Unrecognized test scenario: {test_scenario}")
